+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- fix importing components when one file is a prefix of the other in the same directory
+
 ## [14.0.2] - 2019-03-10
 
 - prevent `bit init` from initialize a non-empty scope when `.bitmap` was deleted unless `--force` is used